requirements.txt
mmcv-full>=1.3.14,<1.4
mmdet>=2.18,<2.19
nncore>=0.3.2
torch>=1.10
torchvision>=0.11
